fix: merge main, resolve conflicts, fix verification module
- Restore symphonia dependency for elevenlabs audio decoding
- Remove duplicate determine.rs (replaced by scoring.rs)
- Fix tests to import from scoring module
- Make scoring constants pub(crate) for test access
