+++ .gitignore
@@ -14,4 +14,5 @@ analysis/optimization/fp-sweep-runs/
 /timeline.json
 /verified.json
 /*.wav
-__pycache__
+.mimocode/
+__pycache__/
