+++ crates/movie-radio-verification/src/lib.rs
@@ -2,5 +2,5 @@ pub mod verification;
 
 pub use verification::{
     default_filter_segment_confidence_ceiling, filter_low_confidence_non_voice_segments,
-    verify_timeline, VerificationReport, VerificationStatus,
+    verify_timeline, AppliedThresholds, VerificationReport, VerificationStatus,
 };
