+++ crates/movie-radio-verification/src/verification/scoring.rs
@@ -1,9 +1,49 @@
-use super::{AppliedThresholds, SpectralFeatures, VerificationStatus};
+use crate::verification::analysis::{SpectralFeatures, VerificationStatus};
+use serde::{Deserialize, Serialize};
 
-const SPEECH_ZCR_MIN: f32 = 0.02;
-const SPEECH_ZCR_MAX: f32 = 0.35;
-const VERIFICATION_HIGH_CONFIDENCE_THRESHOLD: f32 = 0.55;
-const GRAPH_STRUCTURE_WEIGHT: f32 = 0.25;
+#[derive(Debug, Clone, Serialize, Deserialize)]
+pub struct AppliedThresholds {
+    pub entropy_min: f32,
+    pub entropy_max: f32,
+    pub flatness_max: f32,
+    pub energy_min: f32,
+    pub centroid_min: f32,
+    pub centroid_max: f32,
+}
+
+pub(crate) const DEFAULT_ENTROPY_MIN: f32 = 3.5;
+pub(crate) const DEFAULT_ENTROPY_MAX: f32 = 7.0;
+pub(crate) const DEFAULT_FLATNESS_MAX: f32 = 0.45;
+pub(crate) const DEFAULT_ENERGY_MIN: f32 = 0.001;
+pub(crate) const DEFAULT_CENTROID_MIN: f32 = 100.0;
+pub(crate) const DEFAULT_CENTROID_MAX: f32 = 6000.0;
+pub(crate) const VERIFICATION_HIGH_CONFIDENCE_THRESHOLD: f32 = 0.55;
+pub(crate) const SPEECH_ZCR_MIN: f32 = 0.02;
+pub(crate) const SPEECH_ZCR_MAX: f32 = 0.35;
+pub(crate) const GRAPH_STRUCTURE_WEIGHT: f32 = 0.25;
+const DEFAULT_FILTER_SEGMENT_CONFIDENCE_CEILING: f32 = 0.55;
+
+pub fn default_filter_segment_confidence_ceiling() -> f32 {
+    DEFAULT_FILTER_SEGMENT_CONFIDENCE_CEILING
+}
+
+pub(crate) fn build_thresholds(
+    entropy_min: Option<f32>,
+    entropy_max: Option<f32>,
+    flatness_max: Option<f32>,
+    energy_min: Option<f32>,
+    centroid_min: Option<f32>,
+    centroid_max: Option<f32>,
+) -> AppliedThresholds {
+    AppliedThresholds {
+        entropy_min: entropy_min.unwrap_or(DEFAULT_ENTROPY_MIN),
+        entropy_max: entropy_max.unwrap_or(DEFAULT_ENTROPY_MAX),
+        flatness_max: flatness_max.unwrap_or(DEFAULT_FLATNESS_MAX),
+        energy_min: energy_min.unwrap_or(DEFAULT_ENERGY_MIN),
+        centroid_min: centroid_min.unwrap_or(DEFAULT_CENTROID_MIN),
+        centroid_max: centroid_max.unwrap_or(DEFAULT_CENTROID_MAX),
+    }
+}
 
 pub(crate) fn determine_verification_status(
     features: &SpectralFeatures,
@@ -86,4 +126,16 @@ fn graph_structure_nonvoice_confidence(
         / 4.0;
 
     ((nonvoice_region_score * 0.7) + ((1.0 - speech_region_score) * 0.3)).clamp(0.0, 1.0)
+}
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+
+    #[test]
+    fn build_thresholds_uses_defaults() {
+        let thresholds = build_thresholds(None, None, None, None, None, None);
+        assert_eq!(thresholds.entropy_min, DEFAULT_ENTROPY_MIN);
+        assert_eq!(thresholds.centroid_max, DEFAULT_CENTROID_MAX);
+    }
 }
