+++ crates/movie-radio-voice/Cargo.toml
@@ -15,6 +15,9 @@ tracing.workspace = true
 ort.workspace = true
 reqwest.workspace = true
 symphonia.workspace = true
+llama-cpp-2 = "0.1"
+once_cell = "1"
+rand = "0.10"
 
 [lints]
 workspace = true
